Added First Project
I'm going to want to turn this into a Library. To figure out how to best do that, I'm going to have to write a few NNs first, just to see how to best abstract things out. This means, I need to add a very first project. The NN course I'm taking is in Python and uses NumPy, so I've already got started with a Shape struct that covers the needs of the NNs I've studied so far.