Made some changes to Shape.
I wanted to increase performance of setting the shape values with the index operator, but that lead to the discover of a few issues. An call to an empty constructor will be used for a struct even when the decalred constructor falls within the use of an empty construtor. Since I can't define an empty constructor for a struct, that means that the array containing the shape doens't get created. So, I switched the Shape from a struct to a class. I really wanted it as a shape, but I want good performance indexing more. Besides, with a struct, the rows and columns would have been initialzed to zero on the default constructor, and that's something I really, really don't want at this time.